+++ CHANGELOG.md
@@ -4,15 +4,24 @@ All notable changes to this project will be documented in this file.
 
 ## Unreleased
 
+### Changed
+
+- simplified how the router work.
+
 ## 0.0.2 - 2023-12-28
+
 ### Added
+
 - `toHref` method allowing the creation of a valid `href` from a path or a `NamedDestination`.
 
 ### Fixed
+
 - improved `navigate()` with the existance of a `base`.
 
 ## 0.0.1 - 2023-12-08
+
 ### Added
+
 - export `RouterInstance` class.
 - export `types`.
 - export `isUrlNavigatable` method.